more comments from pr #82
